Move impala compile tests to backends/impala
